Moved framework detection logic to its own module
